Make space for each modifier as separate module
